# openapi.yaml
openapi: 3.1.0
info:
  title: E-commerce Polyglot API
  version: 1.0.0
  description: API para plataforma de comercio electrónico con persistencia poliglota
  contact:
    name: Equipo de Desarrollo
    url: https://github.com/yourusername/ING2_UADE_TPO
servers:
  - url: http://localhost:3000/api
    description: Servidor de desarrollo
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      required:
        - name
        - description
        - sku
        - category
        - price
        - stock
        - images
      properties:
        _id:
          type: string
          format: mongodb-objectid
        name:
          type: string
        description:
          type: string
        sku:
          type: string
        category:
          type: string
        price:
          type: number
          minimum: 0
        stock:
          type: integer
          minimum: 0
          default: 0
        images:
          type: array
          items:
            type: string
        specifications:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PriceHistory:
      type: object
      required:
        - productId
        - price
        - timestamp
      properties:
        productId:
          type: string
          format: mongodb-objectid
        price:
          type: number
          minimum: 0
        timestamp:
          type: string
          format: date-time
        currency:
          type: string
          default: USD
    ProductChange:
      type: object
      required:
        - productId
        - changeType
        - timestamp
      properties:
        productId:
          type: string
          format: mongodb-objectid
        changeType:
          type: string
          enum: [create, update, delete]
        oldValue:
          type: string
        newValue:
          type: string
        timestamp:
          type: string
          format: date-time
...
    ProductAnalytics:
      type: object
      properties:
        averagePrice:
          type: number
        volatility:
          type: object
          properties:
            changes:
              type: integer
            volatility:
              type: number
        period:
          type: object
          properties:
            start:
              type: string
              format: date-time
            end:
              type: string
              format: date-time
...
paths:
...
  /products:
    get:
      tags: [Products]
      summary: Listar productos
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        200:
          description: Lista paginada de productos
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  total:
                    type: integer
                  page:
                    type: integer
                  totalPages:
                    type: integer
                  hasMore:
                    type: boolean

    post:
      tags: [Products]
      summary: Crear un nuevo producto
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        201:
          description: Producto creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        401:
          description: No autorizado
        403:
          description: No tiene permisos de administrador
  /products/{id}:
    get:
      tags: [Products]
      summary: Obtener detalles de un producto
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: Detalles del producto
          content:
            application/json:
              schema:
                type: object
                properties:
                  product:
                    $ref: '#/components/schemas/Product'
                  priceHistory:
                    type: array
                    items:
                      $ref: '#/components/schemas/PriceHistory'
                  changes:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProductChange'
        404:
          description: Producto no encontrado

    put:
      tags: [Products]
      summary: Actualizar un producto
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        200:
          description: Producto actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        401:
          description: No autorizado
        403:
          description: No tiene permisos de administrador
        404:
          description: Producto no encontrado

    delete:
      tags: [Products]
      summary: Eliminar un producto
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: Producto eliminado exitosamente
        401:
          description: No autorizado
        403:
          description: No tiene permisos de administrador
        404:
          description: Producto no encontrado
  /products/{id}/analytics:
    get:
      tags: [Products]
      summary: Obtener análisis de precios de un producto
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: startDate
          schema:
            type: string
            format: date-time
        - in: query
          name: endDate
          schema:
            type: string
            format: date-time
      responses:
        200:
          description: Análisis de precios del producto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductAnalytics'
        401:
          description: No autorizado
        403:
          description: No tiene permisos de administrador
        404:
          description: Producto no encontrado
...